refactor: split agents into modules and add base interface
