+++ agentic-project-assistant/no_framework/tests/evaluation_agent.py
@@ -3,7 +3,8 @@ import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from agents.base_agents import EvaluationAgent, KnowledgeAugmentedPromptAgent
+from agents.evaluation_agent import EvaluationAgent
+from agents.knowledge_augmented_prompt_agent import KnowledgeAugmentedPromptAgent
 from agents.openai_service import OpenAIService
 from config import load_openai_api_key, load_openai_base_url
 import logging
