+++ agentic-project-assistant/no_framework/tests/knowledge_augmented_prompt_agent.py
@@ -2,7 +2,7 @@ import sys
 import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from agents.base_agents import KnowledgeAugmentedPromptAgent
+from agents.knowledge_augmented_prompt_agent import KnowledgeAugmentedPromptAgent
 from agents.openai_service import OpenAIService
 from config import load_openai_api_key, load_openai_base_url
